requirements.txt
azure-core==1.29.4
azure-storage-blob==12.18.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
idna==3.4
isodate==0.6.1
pycparser==2.21
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
